pyproject.toml
[project]
name = "ml-nids-testbed"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dawetmaster", email = "andika.naufal48@gmail.com"},
]
dependencies = [
    "scapy>=2.5.0",
    "ntplib>=0.4.0",
    "grpcio>=1.64.1",
    "grpcio-tools>=1.64.1",
    "protobuf>=5.27.0",
    "psutil>=5.9.8",
    "numpy>=2.0.0",
    "fake-useragent>=1.5.1",
    "pyyaml>=6.0.1",
    "uvicorn>=0.30.3",
    "fastapi>=0.111.1",
    "requests>=2.32.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[tool.pdm.scripts]
agent = "python src/agent/agent.py"
controller = "python src/controller/main.py"
